requirements.txt
torch>=2.0.1
numpy>=1.21.2
transformers>=4.38.2
timm>=0.9.16
accelerate
sentencepiece
attrdict
einops
opencv-python
spaces
matplotlib
pillow>=11.1.0

# for gradio demo
# gradio==3.48.0
# gradio-client==0.6.1
# mdtex2html==1.3.0
# pypinyin==0.50.0
# tiktoken==0.5.2
# tqdm==4.64.0
# colorama==0.4.5
# Pygments==2.12.0
# markdown==3.4.1
# SentencePiece==0.1.96
